refactor: split Map to Container and Presenter
