+++ src/MapContainer.js
@@ -2,65 +2,9 @@ import React, { useEffect, useState } from "react";
 import GetPosition from "./Getloc";
 import { CreateInfo } from "./CreateInfo";
 import { maskInfo } from "./api";
-import Display, { sortingStore } from "./DisplayStore";
-import styled from "styled-components";
+import MapPresenter from "./MapPresenter";
+import { sortingStore } from "./DisplayStore";
 
-const DisplayWrapper = styled.div`
-  position: relative;
-  margin-left: 5px;
-`;
-const DisplayList = styled.div`
-  position: fixed;
-  width: 90vw;
-  overflow: auto;
-  height: 33vh;
-  margin-top: 3px;
-  margin-left: 5px;
-  padding-left: 8px;
-`;
-const Column = styled.div`
-  margin-left: 10px;
-  padding-bottom: 5px;
-  padding-right: 10px;
-  padding-left: 8px;
-  display: grid;
-  grid-template-columns: 2fr 2fr 0.9fr;
-  width: 90%;
-  font-weight: bold;
-  font-size: 0.87em;
-  background-color: #ced6e0;
-  box-shadow: 1.5px 1.5px;
-`;
-const Toolbar = styled.button`
-  position: fixed;
-  z-index: 9999;
-  right: 9%;
-  bottom: 46%;
-  padding: 5px 7px 5px 7px;
-  font-size: 0.9em;
-  font-weight: bold;
-  background-color: ${props => (props.isStock ? "#ffeaa7" : "#ffffff")};
-  border-radius: 15px;
-  outline: none;
-`;
-const InitLoc = styled.button`
-  position: fixed;
-  z-index: 9999;
-  right: 9%;
-  bottom: 52.5%;
-  padding: 5px 7px 5px 7px;
-  font-size: 0.9em;
-  font-weight: bold;
-  background-color: #ffffff;
-  border-radius: 15px;
-  outline: none;
-`;
-const Updated = styled.div`
-  position: fixed;
-  right: 3%;
-  top: 1.2%;
-  font-weight: bold;
-`;
 const storeMarkers = [];
 const SetMap = () => {
   const [map, setMymap] = useState(null);
@@ -82,11 +26,11 @@ const SetMap = () => {
       clearMap();
       try {
         const {
-          coords: { latitude, longitude }
+          coords: { latitude, longitude },
         } = await GetPosition();
         setPosition({ latitude, longitude });
         let {
-          data: { stores }
+          data: { stores },
         } = await maskInfo.storesByGeo(latitude, longitude);
 
         setStore(sortingStore(stores, position));
@@ -95,7 +39,7 @@ const SetMap = () => {
         const container = document.getElementById("map"),
           options = {
             center: new window.kakao.maps.LatLng(latitude, longitude),
-            level: 4
+            level: 4,
           };
         if (map) {
           map.panTo(options.center);
@@ -120,7 +64,7 @@ const SetMap = () => {
     // show Current Location
     const markerPosition = new window.kakao.maps.LatLng(latitude, longitude);
     const markers = new window.kakao.maps.Marker({
-      position: markerPosition
+      position: markerPosition,
     });
     if (marker) marker.setMap(null);
     markers.setMap(map);
@@ -130,11 +74,11 @@ const SetMap = () => {
   //get store info
   useEffect(() => {
     if (stores != null) {
-      stores.map(data => {
+      stores.map((data) => {
         const { lat, lng, name, remain_stat } = data;
         const storemarker = CreateInfo(map, lat, lng, {
           name,
-          remain_stat
+          remain_stat,
         });
         storeMarkers.push(storemarker);
         return "";
@@ -158,45 +102,16 @@ const SetMap = () => {
   }, [isStock, map, stores]);
 
   return (
-    <div>
-      <Updated>
-        업데이트:{" "}
-        {lastUpdated ? lastUpdated.slice(5, lastUpdated.length) : null}
-      </Updated>
-      <Toolbar
-        isStock={isStock}
-        onClick={() => {
-          setStock(isStock ? false : true);
-        }}
-      >
-        재고있음
-      </Toolbar>
-      <InitLoc
-        onClick={() => {
-          setToggle(!toggle);
-        }}
-      >
-        내위치
-      </InitLoc>
-      <DisplayWrapper>
-        <Column>
-          <div>약국명칭</div>
-          <div>입고시간</div>
-          <div>재고현황</div>
-        </Column>
-        <DisplayList>
-          {!stores ? (
-            "Load"
-          ) : (
-            <Display
-              stores={sortingStore(stores, position)}
-              isStock={isStock}
-              map={map}
-            ></Display>
-          )}
-        </DisplayList>
-      </DisplayWrapper>
-    </div>
+    <MapPresenter
+      lastUpdated={lastUpdated}
+      isStock={isStock}
+      toggle={toggle}
+      stores={stores}
+      position={position}
+      map={map}
+      setToggle={setToggle}
+      setStock={setStock}
+    ></MapPresenter>
   );
 };
